third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/css-live-edit-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that styles are updated when live-editing css resource.


Running: testLiveEdit
display: block;
    block - div user agent stylesheet
font-size: 20px;
    20px - body css-live-edit.css:1 -> css-live-edit.css:1:7
[expanded] 
element.style { ()
